requirements.txt
pdfminer.six
python-docx
spacy

regex
nltk
